Move filtering to model directory since used in components
