+++ client/src/store/historyStore/historyItemsStore.js
@@ -7,7 +7,7 @@ import { reverse } from "lodash";
 import { LastQueue } from "utils/promise-queue";
 import { urlData } from "utils/url";
 import { mergeArray } from "./model/utilities";
-import { getFilters, getQueryDict, testFilters } from "./historyItemsFiltering";
+import { getFilters, getQueryDict, testFilters } from "./model/filtering";
 
 const limit = 100;
 const queue = new LastQueue();
